refactor: move GuiItem to items package
